+++ frontend/app/player/web/MessageManager.ts
@@ -221,7 +221,7 @@ export default class MessageManager {
         fileReader.append(b)
         const msgs: Array<Message> = []
         for (let msg = fileReader.readNext();msg !== null;msg = fileReader.readNext()) {
-          msgs.push(msg)
+          msg && msgs.push(msg)
         }
         const sorted = msgs.sort((m1, m2) => {
           // @ts-ignore