+++ tracker/tracker/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Capture mouse thrashing, input hesitation+duration, click hesitation
 - Capture DOM node drop event (>30% nodes removed)
 - Capture iframe network requests
+- added `{ network: { disableClickmaps: boolean } }` to disable calculating el. selectors
 
 ## 5.0.1